- (*): Merged the MT changes into the main branch.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@1662 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/asn1.c
@@ -433,7 +433,7 @@ asn_parse_header(data, datalength, type)
       if (bufp == NULL)
         return NULL;
       header_len = bufp - data;
-      if (header_len + asn_length > *datalength){
+      if ((header_len + (int)asn_length) > *datalength){
         ERROR_MSG("asn length too long");
         return NULL;
       }
@@ -986,7 +986,7 @@ asn_parse_unsigned_int64(data, datalength, type, cp, countersize)
 	bufp = asn_parse_length(bufp + 2, &asn_length);
         if (bufp == NULL)
             return NULL;
-        if (asn_length + (bufp - data) > *datalength){
+        if ((int)(asn_length + (bufp - data)) > *datalength){
             ERROR_MSG("overflow of message");
             return NULL;
         }
@@ -1163,7 +1163,7 @@ asn_parse_signed_int64(data, datalength, type, cp, countersize)
     bufp = asn_parse_length(bufp + 2, &asn_length);
     if (bufp == NULL)
       return NULL;
-    if (asn_length + (bufp - data) > *datalength){
+    if ((int)(asn_length + (bufp - data)) > *datalength){
       ERROR_MSG("overflow of message");
       return NULL;
     }
@@ -1291,7 +1291,7 @@ asn_parse_float(data, datalength, type, floatp, floatsize)
 	ERROR_MSG("bad length");
 	return NULL;
     }
-    if (asn_length + (bufp - data) > *datalength){
+    if ((int)(asn_length + (bufp - data)) > *datalength){
 	ERROR_MSG("overflow of message");
 	return NULL;
     }
@@ -1304,7 +1304,7 @@ asn_parse_float(data, datalength, type, floatp, floatsize)
 	bufp = asn_parse_length(bufp + 2, &asn_length);
         if (bufp == NULL)
             return NULL;
-        if (asn_length + (bufp - data) > *datalength){
+        if ((int)(asn_length + (bufp - data)) > *datalength){
             ERROR_MSG("overflow of message");
             return NULL;
         }
@@ -1410,7 +1410,7 @@ asn_parse_double(data, datalength, type, doublep, doublesize)
 	ERROR_MSG("bad length");
 	return NULL;
     }
-    if (asn_length + (bufp - data) > *datalength){
+    if ((int)(asn_length + (bufp - data)) > *datalength){
 	ERROR_MSG("overflow of message");
 	return NULL;
     }
@@ -1423,7 +1423,7 @@ asn_parse_double(data, datalength, type, doublep, doublesize)
 	bufp = asn_parse_length(bufp + 2, &asn_length);
         if (bufp == NULL)
             return NULL;
-        if (asn_length + (bufp - data) > *datalength){
+        if ((int)(asn_length + (bufp - data)) > *datalength){
             ERROR_MSG("overflow of message");
             return NULL;
         }
